LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_put8-expected.txt


PASS IDBObjectStore.put() - object store has autoIncrement:true and the key path is an object attribute
